CMS-203: Change file name natureAndCulture to about
